Prune the deprecated @Health qualifier and remove all references of it
