+++ spec/src/main/asciidoc/java-api.adoc
@@ -42,7 +42,6 @@ The nature of the procedure is defined by annotating the `HealthCheck` procedure
 
 * Readiness checks defined with `@Readiness` annotation
 * Liveness checks defined with `@Liveness` annotation
-* Backward compatible checks defined with `@Health` annotation
 
 
 A `HealthCheck` procedure with none of the above annotations is not an active procedure and should be ignored.
@@ -59,13 +58,6 @@ A Health Check for liveness allows third party services to determine if the appl
 This means that if this procedure fails the application can be discarded (terminated, shutdown).
 
 The `@Liveness` annotation must be applied on a `HealthCheck` implementation to define a Liveness check procedure, otherwise, this annotation is ignored.
-
-
-=== Backward compatible check
-
-To provide backward compatibility with previous specification version, a `HealthCheck` implementation with `@Health` annotation is still supported.
-
-`@Health` annotation is deprecated, new procedures shouldn't use it.
 
 == Multiple HealthChecks procedures for a given kind
 
@@ -149,7 +141,7 @@ public class CheckDiskspace implements HealthCheck {
 
 == Integration with CDI
 
-Any enabled bean with a bean of type `org.eclipse.microprofile.health.HealthCheck` and `@Liveness`, `@Readiness` or `@Health` qualifier can be used as health check procedure.
+Any enabled bean with a bean of type `org.eclipse.microprofile.health.HealthCheck` and `@Liveness` or `@Readiness` qualifier can be used as health check procedure.
 
 
 Contextual references of health check procedures are invoked by runtime when the outermost protocol entry point (i.e. `http://HOST:PORT/health`) receives an inbound request
