--- tck/src/main/java/org/eclipse/microprofile/health/tck/deployment/SuccessfulHealth.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2017 Contributors to the Eclipse Foundation
- *
- * See the NOTICES file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * You may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- */
-package org.eclipse.microprofile.health.tck.deployment;
-
-import org.eclipse.microprofile.health.Health;
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-
-import javax.enterprise.context.ApplicationScoped;
-
-/**
- * @author Heiko Braun
- * @since 13.06.17
- */
-@Health
-@ApplicationScoped
-public class SuccessfulHealth implements HealthCheck {
-    @Override
-    public HealthCheckResponse call() {
-        return HealthCheckResponse.up("successful-check");
-    }
-}
